Use MockTransportHandler in tests
