+++ tests/conftest.py
@@ -21,3 +21,15 @@ def sampled_zipkin_attr(zipkin_attributes):
 @pytest.fixture
 def unsampled_zipkin_attr(zipkin_attributes):
     return ZipkinAttrs(is_sampled=False, **zipkin_attributes)
+
+
+class MockTransportHandler(object):
+
+    def __init__(self, max_payload_bytes=None):
+        self.max_payload_bytes = max_payload_bytes
+
+    def __call__(self, *args, **kwargs):
+        return args[0]
+
+    def get_max_payload_bytes(self):
+        return self.max_payload_bytes
